+++ app/smartbrain_3000/web/_app/immutable/chunks/D5N1T76W.js
@@ -1,0 +1,1 @@
+import"./CoflnhAA.js";