+++ app/smartbrain_3000/web/_app/immutable/entry/start.B5jEToGv.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/CoflnhAA.js";export{e as load_css,t as start};