+++ app/smartbrain_3000/web/_app/immutable/entry/start.Bs4ek6gx.js
@@ -1,1 +1,0 @@
-import{a as e,n as t}from"../chunks/C8eJsAyG.js";export{e as load_css,t as start};